requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
jupyter
azureml-sdk
azure-storage-blob
azureml-defaults
fastapi
joblib
numpy
uvicorn
azure-storage-blob==12.19.1